moved function to backend folder
